backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
python-multipart
langchain 
langchain-openai
langchain-ollama 
langchain-chroma
langchain-core
langchain-community
langchain-nomic
ollama
pandas
numpy
scikit-learn
requests
newsapi-python
pdfplumber
beautifulsoup4
supabase
openai
linkup-sdk
mermaid-py
Pillow
